improve performance

+++ include/global.hpp
@@ -1,5 +1,5 @@
 #include <stddef.h>
 
-constexpr const float sampling_rate = 200000;
+constexpr const uint64_t sampling_rate = 200000;
 constexpr const size_t num_channels = 4;
